Renamed post-processors, refactored reco interaction attributes
