+++ spine/post/reco/__init__.py
@@ -13,3 +13,4 @@ from .pid import *
 from .kinematics import *
 from .label import *
 from .shower import *
+from .topology import *
