+++ spine/post/reco/shower.py
@@ -4,21 +4,18 @@ from spine.utils.globals import (PHOT_PID, PROT_PID, PION_PID, ELEC_PID,
                                  SHOWR_SHP, TRACK_SHP)
 
 from spine.post.base import PostBase
-from scipy.spatial.distance import cdist
 from sklearn.cluster import DBSCAN
 from sklearn.metrics.pairwise import cosine_similarity
 
-from sklearn.decomposition import PCA
-
-from spine.utils.gnn.cluster import cluster_dedx_dir, cluster_dedx_DBScan_PCA, cluster_dedx
 from spine.utils.tracking import get_track_segment_dedxs
+
+from spine.utils.numba_local import cdist
 
 from scipy.stats import pearsonr
 from sklearn.cluster import DBSCAN
 
-__all__ = ['ConversionDistanceProcessor', 'ShowerMultiArmCheck', 
-           'ShowerStartpointCorrectionProcessor', 'ShowerdEdXProcessor',
-           'ShowerSpreadProcessor', 'ShowerTrunkValidityProcessor',
+__all__ = ['ConversionDistanceProcessor', 
+           'ShowerStartpointCorrectionProcessor', 
            'MichelTaggingProcessor']
 
 
@@ -122,10 +119,8 @@ class ConversionDistanceProcessor(PostBase):
                             
             if leading_shower is None:
                 ia.leading_shower_vertex_distance = -np.inf
-                ia.leading_shower_num_fragments = -1
             else:
                 ia.leading_shower_vertex_distance = leading_shower.vertex_distance
-                ia.leading_shower_num_fragments = leading_shower.num_fragments
             
     @staticmethod        
     def convdist_protons(ia, shower_p):
@@ -242,89 +237,6 @@ class ConversionDistanceProcessor(PostBase):
             conversion_dist = min(sep, conversion_dist)
 
         return conversion_dist
-    
-    
-class ShowerMultiArmCheck(PostBase):
-    """Check whether given primary electron candidate is likely
-    to be a merged multi-particle shower.
-    
-    This processor computes direction vectors of the shower points
-    from the shower start and performs DBSCAN clustering on the unit sphere
-    using the cosine distance metric. If there are more than one cluster that
-    has a mean direction vector outside a certain angular threshold, the
-    shower is considered to be a multi-arm shower and is rejected as 
-    a valid primary electron candidate.
-    
-    NOTE: This processor can only change reco electron shower pid to
-    photon pid depending on the angle threshold. 
-    """
-
-    # Name of the post-processor (as specified in the configuration)
-    name = 'shower_multi_arm_check'
-
-    # Alternative allowed names of the post-processor
-    aliases = ('shower_multi_arm',)
-    
-    def __init__(self, threshold=70, min_samples=20, eps=0.02, inplace=True):
-        """Specify the threshold for the number of arms of showers.
-
-        Parameters
-        ----------
-        threshold : float, default 70 (deg)
-            If the electron shower's leading and subleading angle are
-            separated by more than this, the shower is considered to be
-            invalid and its PID will be changed to PHOT_PID.
-        min_samples : int, default 20
-            The number of samples (or total weight) in a neighborhood 
-            for a point to be considered as a core point (DBSCAN).
-        eps : float, default 0.02
-            Maximum distance between two samples for one to be considered
-            as in the neighborhood of the other (DBSCAN).
-        """
-        super().__init__('interaction', 'reco')
-        
-        self.threshold = threshold
-        self.min_samples = min_samples
-        self.eps = eps
-        self.inplace = inplace
-        
-    def process(self, data):
-        """Update reco interaction topologies using the shower multi-arm check.
-
-        Parameters
-        ----------
-        data : dict
-            Dictionaries of data products
-        """
-        # Loop over the reco interactions
-        for ia in data['reco_interactions']:
-            # Loop over particles, select the ones that pass a threshold
-            
-            leading_shower, energy = None, -np.inf
-            
-            for p in ia.particles:
-                
-                if p.shape == SHOWR_SHP and p.is_primary:
-                    
-                    angle = self.compute_angular_criterion(p, ia.vertex, 
-                                                     eps=self.eps, 
-                                                     min_samples=self.min_samples)
-                    
-                    p.shower_split_angle = angle
-                    if self.inplace:
-                        if angle > self.threshold:
-                            p.pid = PHOT_PID
-                
-                if p.pid == ELEC_PID and p.is_primary and (p.shape == SHOWR_SHP):
-
-                    if self.inplace:
-                        if angle > self.threshold:
-                            p.pid = PHOT_PID
-            
-            if leading_shower is None:
-                ia.shower_split_angle = -np.inf
-            else:
-                ia.shower_split_angle = leading_shower.shower_split_angle
             
                 
     # @staticmethod
@@ -445,7 +357,6 @@ class ShowerStartpointCorrectionProcessor(PostBase):
         """
         # Loop over the reco interactions
         for ia in data['reco_interactions']:
-            vertex = ia.vertex
             for p in ia.particles:
                 if (p.shape == SHOWR_SHP) and (p.is_primary):
                     new_point = self.correct_shower_startpoint(p, ia)
@@ -481,312 +392,6 @@ class ShowerStartpointCorrectionProcessor(PostBase):
             return shower_p.start_point
         guess = shower_p.points[closest_idx[0]]
         return guess
-
-
-class ShowerdEdXProcessor(PostBase):
-    """Compute the dEdX of the primary EM shower
-    by summing the energy depositions along the shower trunk and dividing
-    by the total length of the trunk.
-    """
-    
-    name = 'shower_dedx_processor'
-    aliases = ('shower_dedx',)
-    
-    def __init__(self, threshold=4.0, max_dist=3.0, mode='direction', inplace=True):
-        """Specify the EM shower dEdX threshold.
-
-        Parameters
-        ----------
-        threshold : float, default 4.0
-            If the dEdX of the shower is greater than this, the shower
-            will be considered a photon.
-        max_dist : float, default 3.0
-            Maximum distance to consider for the dEdX calculation.
-        mode : str, default 'direction'
-            Method to use for dEdX calculation.
-        inplace : bool, default True
-            If True, the processor will update the reco interaction in-place.
-        """
-        super().__init__('interaction', 'reco')
-        self.threshold = threshold
-        self.max_dist = max_dist
-        self.mode = mode
-        self.inplace = inplace
-        
-    def process(self, data):
-        """Compute the shower dEdX and modify the PID if inplace=True.
-
-        Parameters
-        ----------
-        data : dict
-            Dictionaries of data products
-        """
-        # Loop over the reco interactions
-        for ia in data['reco_interactions']:
-            
-            leading_shower, max_ke = None, -np.inf
-            
-            for p in ia.particles:
-                if (p.shape == SHOWR_SHP) and (p.is_primary):
-                    
-                    if self.mode == 'legacy':
-                        dedx = cluster_dedx(p.points, p.depositions, p.start_point, max_dist=self.max_dist)
-                    elif self.mode == 'dbscan':
-                        dedx = cluster_dedx_DBScan_PCA(p.points, p.depositions, p.start_point, 
-                                                    self.max_dist, simple=True)
-                    elif self.mode == 'direction':
-                        dedx = cluster_dedx_dir(p.points,
-                                                p.depositions,
-                                                p.start_point,
-                                                p.start_dir,
-                                                dedx_dist=self.max_dist,
-                                                simple=True)
-                    else:
-                        raise ValueError('Invalid dEdX calculation mode')
-                    p.shower_dedx = dedx
-                    
-                    if p.ke > max_ke:
-                        leading_shower = p
-                        max_ke = p.ke
-                    
-                    if self.inplace:
-                        if dedx >= self.threshold:
-                            p.pid = PHOT_PID
-            
-            if leading_shower is not None:
-                ia.leading_shower_dedx = leading_shower.shower_dedx
-            else:
-                ia.leading_shower_dedx = -1.
-            
-            
-class ShowerSpreadProcessor(PostBase):
-    """Compute the spread of the primary EM shower
-    by computing the mean direction and the weighted average cosine distance
-    with respect to the mean direction.
-    """
-    
-    name = 'shower_spread_processor'
-    aliases = ('shower_spread',)
-    
-    def __init__(self, threshold=0.043, length_scale=14.0, 
-                 refvox_mode='vertex', inplace=True):
-        """Specify the EM shower spread threshold.
-
-        Parameters
-        ----------
-        threshold : float, default 4.0
-            If the spread of the shower is greater than this, the shower
-            will be considered a photon.
-        inplace : bool, default True
-            If True, the processor will update the reco interaction in-place.
-        """
-        super().__init__('interaction', 'reco')
-        self.threshold = threshold
-        self.length_scale = length_scale
-        self.refvox_mode = refvox_mode
-        self.inplace = inplace
-        
-    def process(self, data):
-        """Compute the shower spread and modify the PID if inplace=True.
-
-        Parameters
-        ----------
-        data : dict
-            Dictionaries of data products
-        """
-        # Loop over the reco interactions
-        for ia in data['reco_interactions']:
-            
-            leading_shower, max_ke = None, -np.inf
-            
-            for p in ia.particles:
-                
-                if p.is_primary:
-                    p.axial_pearsonr = compute_axial_pearsonr(p)
-                
-                if (p.shape == SHOWR_SHP) and (p.is_primary):
-                    
-                    if self.refvox_mode == 'vertex':
-                    
-                        spread = compute_shower_spread(p.points,
-                                                    ia.vertex,
-                                                    l=self.length_scale)
-                    
-                    elif self.refvox_mode == 'startpoint':
-                        
-                        spread = compute_shower_spread(p.points,
-                                                    p.start_point,
-                                                    l=self.length_scale)
-                        
-                    else:
-                        raise ValueError('Invalid reference voxel mode')
-                    
-                    p.shower_spread = spread
-                    
-                    if p.ke > max_ke:
-                        leading_shower = p
-                        max_ke = p.ke
-                    
-                    if self.inplace:
-                        if spread >= self.threshold:
-                            p.pid = PHOT_PID
-            
-            if leading_shower is not None:
-                ia.leading_shower_spread = leading_shower.shower_spread
-                ia.leading_shower_axial_pearsonr = leading_shower.axial_pearsonr
-            else:
-                ia.leading_shower_spread = -1.
-                ia.leading_shower_axial_pearsonr = -1.
-    
-    
-def compute_axial_pearsonr(shower_p):
-    """Compute the pearson R correlation coefficient between the
-    distance of the shower points from the startpoint along the shower
-    axis and the perpendicular distance from the shower axis.
-
-    Parameters
-    ----------
-    shower_p : RecoParticle
-        Primary EM shower to compute the axial pearson R.
-
-    Returns
-    -------
-    The pearson R correlation coefficient between (-1, 1)
-    """
-    
-    if len(shower_p.points) < 3:
-        return -1.
-    
-    startpoint = shower_p.start_point
-    v0 = shower_p.start_dir
-    
-    dists = np.linalg.norm(shower_p.points - startpoint, axis=1)
-    v = (startpoint - shower_p.points) - np.sum((startpoint - shower_p.points) * v0, axis=1, keepdims=True) \
-      * np.broadcast_to(v0, shower_p.points.shape)
-    perps = np.linalg.norm(v, axis=1)
-    
-    out = pearsonr(dists, perps)
-    return out[0]
-
-
-def compute_shower_spread(points, vertex, l=14.0):
-    """Compute the spread of the shower by computing the mean direction and
-    the weighted average cosine distance with respect to the mean direction.
-    
-    Parameters
-    ----------
-    points : np.ndarray
-        (N, 3) array of the shower points.
-    l : float, default 14.0
-        Length scale for the exponential weighting.
-        
-    Returns
-    -------
-
-    spread : float
-        Spread cut parameter of the shower.
-    """
-    dists = np.linalg.norm(points - vertex, axis=1)
-    mask = dists > 0
-    if mask.sum() == 0:
-        return -np.inf
-
-    # Compute Spread (whole shower)
-    directions = (points[mask] - vertex) / dists[mask].reshape(-1, 1)
-    weights = np.clip(np.exp(- dists[mask] / l), min=1e-6)
-    mean_direction = np.average(directions, weights=weights, axis=0)
-    mean_direction /= np.linalg.norm(mean_direction)
-    cosine = 1 - np.sum(directions * mean_direction.reshape(1, -1), axis=1)
-    spread = np.average(cosine, weights=weights)
-
-    return spread
-
-
-class ShowerTrunkValidityProcessor(PostBase):
-    """Compute the validity of the shower trunk by computing the PCA
-    principal explained variance ratio of the shower points.
-    """
-    
-    name = 'shower_trunk_validity'
-    aliases = ('shower_trunk_processor',)
-    
-    def __init__(self, threshold, inplace=True):
-        """Specify the EM shower spread thresholds.
-
-        Parameters
-        ----------
-        threshold : float
-            If the spread of the shower trunk is less than this value, the
-            particle is labeled as invalid.
-        inplace : bool, default True
-            If True, the processor will update the reco interaction in-place.
-        """
-        super().__init__('interaction', 'reco')
-        self.threshold = threshold
-        self.inplace = inplace
-        
-    def process(self, data):
-        """Compute the shower spread and modify the PID if inplace=True.
-
-        Parameters
-        ----------
-        data : dict
-            Dictionaries of data products
-        """
-        # Loop over the reco interactions
-        for ia in data['reco_interactions']:
-            
-            leading_shower, max_ke = None, -np.inf
-            
-            for p in ia.particles:
-                
-                trunk_validity = compute_trunk_validity(p)
-                p.trunk_validity = trunk_validity
-                
-                if p.is_primary:
-                    
-                    if p.shape == SHOWR_SHP and p.ke > max_ke:
-                        leading_shower = p
-                        max_ke = p.ke
-                    
-                    if self.inplace:
-                        if trunk_validity < self.threshold:
-                            p.is_valid = False
-                            p.is_primary = False
-            
-            if leading_shower is not None:
-                ia.leading_shower_trunk_validity = leading_shower.trunk_validity
-            else:
-                ia.leading_shower_trunk_validity = -1.
-                
-                
-def compute_trunk_validity(shower_p, r=3.0, n_components=3):
-    """Helper function to compute the validity of the shower trunk
-    by computing the PCA principal explained variance ratio. 
-
-    Parameters
-    ----------
-    shower_p : RecoParticle
-        Primary EM shower to compute the trunk validity. (Doesn't need
-        to be a shower, can be any particle with points)
-    r : float, default 3.0
-        Radius to search for points near the startpoint.
-    n_components : int, default 3
-        Number of components to keep
-
-    Returns
-    -------
-    The first explained variance ratio of the PCA of the shower trunk.
-    """
-    dists = np.linalg.norm(shower_p.points - shower_p.start_point, axis=1)
-    mask = dists < r
-    pts = shower_p.points[mask]
-    if len(pts) <= n_components:
-        return -np.inf
-    else:
-        pca = PCA(n_components=n_components)
-        pca.fit(pts)
-        return pca.explained_variance_ratio_[0]
     
     
 class MichelTaggingProcessor(PostBase):
@@ -829,8 +434,6 @@ class MichelTaggingProcessor(PostBase):
         # Loop over the reco interactions
         for ia in data['reco_interactions']:
             
-            shower_p, max_ke = None, -np.inf
-            
             for p in ia.particles:
                 
                 if p.pid == ELEC_PID and (p.shape == 0 or p.shape == 2):
@@ -838,19 +441,9 @@ class MichelTaggingProcessor(PostBase):
                                                  adj_threshold=self.adj_threshold)
                     if (len(x) > 2) and (len(x) == len(y)):
                         bragg = pearsonr(x, y)[0]
-                        p.adjacent_bragg = bragg
+                        p.adjacent_bragg_pearsonr = bragg
                     else:
-                        p.adjacent_bragg = np.inf
-                        
-                if (p.shape == SHOWR_SHP) and (p.is_primary):
-                    if p.ke > max_ke:
-                        shower_p = p
-                        max_ke = p.ke
-            
-            if shower_p is not None:
-                ia.leading_shower_adjacent_bragg = shower_p.adjacent_bragg
-            else:
-                ia.leading_shower_adjacent_bragg = np.inf  
+                        p.adjacent_bragg_pearsonr = np.inf
                 
             
     def check_bragg_peak(self, shower_p, nu_reco, r, adj_threshold=1.0):
